# contracts/openapi.yaml
openapi: 3.1.0
info:
  title: Momentum Dashboard API
  version: 0.7.0
  summary: OpenAPI-first contracts for React + FastAPI app (Parquet+SQLite).
  description: Contract-first spec. Clients must be generated from this file. All
    endpoints are local-only and authenticated as N/A for now.
servers:
- url: http://localhost:8000/api/v1
...
paths:
...
  /api/v1/screener:
    get:
      tags:
      - Screener
      summary: List screener rows (filtered, sorted, paginated)
      parameters:
      - name: page
        in: query
        schema:
          type: integer
          minimum: 1
          default: 1
      - name: per_page
        in: query
        schema:
          type: integer
          minimum: 1
          maximum: 200
          default: 100
      - name: sort_by
        in: query
        schema:
          type: string
          enum:
          - score
          - pct_today
          - rsi
          - adx
          - proximity_52w_high
          - relvol20
          - price
          - name
          - sector
          - wk_change
          - wk_change_pct
          default: score
      - name: sort_dir
        in: query
        schema:
          type: string
          enum:
          - asc
          - desc
          default: desc
      - name: universe
        in: query
        schema:
          type: string
          enum:
          - NIFTY50
          - NIFTY100
          - NIFTY500
          - MIDCAP
          - SMALLCAP
          - ALL
      - name: symbol
        in: query
        schema:
          type: string
        description: Exact ticker symbol (case-insensitive)
      - name: sector_in
        in: query
        schema:
          type: array
          items:
            type: string
        style: form
        explode: true
      - name: score_min
        in: query
        schema:
          type: number
      - name: rsi_min
        in: query
        schema:
          type: number
      - name: rsi_max
        in: query
        schema:
          type: number
      - name: adx_min
        in: query
        schema:
          type: number
      - name: price_min
        in: query
        schema:
          type: number
      - name: price_max
        in: query
        schema:
          type: number
      - name: pct_today_min
        in: query
        schema:
          type: number
      - name: proximity_52w_high_max
        in: query
        schema:
          type: number
          description: Within X% of 52W high if <= X
      - name: watchlist_only
        in: query
        schema:
          type: boolean
          default: false
      - name: stale_exclude
        in: query
        schema:
          type: boolean
          default: true
      - name: run_id
        in: query
        schema:
          type: string
          description: If omitted, server uses latest snapshot
      responses:
        '200':
          description: Screener result page
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ScreenerList'
  /api/v1/screener/top-movers:
    get:
      tags:
      - Screener
      summary: Top gainers and losers for the requested period
      operationId: getTopMovers
      parameters:
      - name: period
        in: query
        description: Period over which to compute percentage change
        schema:
          type: string
          enum: ["1d", "1w", "1m", "3m"]
          default: "1d"
      responses:
        '200':
          description: Top movers
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TopMovers'
...
components:
...
  schemas:
...
    Pagination:
      type: object
      properties:
        page:
          type: integer
        per_page:
          type: integer
        total:
          type: integer
          nullable: true
        next_cursor:
          type: string
          nullable: true
...
    ScreenerRow:
      type: object
      required:
      - symbol
      - last
      - change_pct
      - score
      properties:
        symbol:
          type: string
          example: RELIANCE
        name:
          type: string
          example: Reliance Industries
        sector:
          type: string
          example: Energy
        last:
          type: number
          format: float
          example: 2700.5
        change_pct:
          type: number
          format: float
          example: 1.2
        wk_change:
          type: number
          format: float
          nullable: true
          example: 35.8
        wk_change_pct:
          type: number
          format: float
          nullable: true
          example: 1.35
        badges:
          type: array
          items:
            $ref: '#/components/schemas/Badge'
        score:
          type: number
          format: float
          example: 78
        strength:
          type: string
          example: Moderate
        rsi:
          type: number
          format: float
          example: 59.32
        adx:
          type: number
          format: float
          example: 17.91
        ret_12_1m:
          type: number
          format: float
          example: 15.4
        ret_6m:
          type: number
          format: float
          example: 10.8
        ret_3m:
          type: number
          format: float
          example: 6.1
        ret_1m:
          type: number
          format: float
          example: 2.3
        pct_from_52w_high:
          type: number
          format: float
          example: -4.2
        atr_pct:
          type: number
          format: float
          example: 2.1
        liquidity:
          type: number
          format: float
          example: 180000000
        vol_spike:
          type: number
          format: float
          example: 1.4
        pct_today:
          type: number
          format: float
          example: 0.85
        buy:
          type: boolean
          example: false
        reason:
          type: string
          example: ''
        source:
          type: string
          example: scan
        stale:
          type: boolean
          example: false
        run_id:
          type: string
          example: 20250910T153000Z
        as_of:
          type: string
          format: date-time
          example: '2025-09-10T15:30:00Z'
        last_index:
          type: string
          format: date
          example: '2025-09-01'
    ScreenerList:
      type: object
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/ScreenerRow'
        pagination:
          $ref: '#/components/schemas/Pagination'
        as_of:
          type: string
          format: date-time
          nullable: true
          description: Snapshot timestamp (UTC)
        run_id:
          type: string
          nullable: true
          description: Run identifier used for this list
    TopMovers:
      type: object
      required: [period, generated_at, gainers, losers]
      properties:
        period:
          type: string
          enum: ["1d", "1w", "1m", "3m"]
        generated_at:
          type: string
          format: date-time
        run_id:
          type: string
          nullable: true
        as_of:
          type: string
          format: date
          nullable: true
        gainers:
          type: array
          items:
            $ref: '#/components/schemas/TopMoverEntry'
        losers:
          type: array
          items:
            $ref: '#/components/schemas/TopMoverEntry'
    TopMoverEntry:
      type: object
      required: [symbol, price, change_pct, next_action]
      properties:
        symbol:
          type: string
        name:
          type: string
          nullable: true
        sector:
          type: string
          nullable: true
        price:
          type: number
          format: double
        change_pct:
          type: number
          format: double
        score:
          type: number
          format: double
          nullable: true
        next_action:
          $ref: '#/components/schemas/DrawerNextAction'
...
    Badge:
      type: object
      required: [category, label]
      properties:
        category:
          type: string
          enum: [BREAKOUT, MOMENTUM, WATCH, IGNORE,ACTION]
        label:
          type: string
          example: "🚀 Very High Breakout"
...
    DrawerNextAction:
      type: object
      required: [code, text, refs]
      properties:
        code:
          type: string
          enum:
            - SELL_NOW
            - SELL_TOMORROW
            - HOLD
            - HOLD_BREAKEVEN
            - HOLD_TIGHT
            - BUY_BREAKOUT
            - BUY_PULLBACK
            - BUY_STARTER
            - WATCH
            - IGNORE
        text:
          type: string
          example: "Hold — Trail stop at ₹4382.0"
        reasons:
          type: array
          items: { type: string }
        refs:
          type: object
          additionalProperties: { type: number, format: double }
          description: Numeric references used in the text (e.g., S/E/pivot/entry/A/B).
...